Added functions for reading from sfall INI files more easily; Refactor Karma module init: use new INI functions, get rid of hacky string split code, removed KarmaFRMsCount (it's redundant)

+++ sfall/Modules/Graphics.cpp
@@ -959,7 +959,7 @@ HRESULT _stdcall FakeDirectDrawCreate2(void*, IDirectDraw** b, void*) {
 	return DD_OK;
 }
 
-static DWORD fadeMulti;
+static double fadeMulti;
 static __declspec(naked) void FadeHook() {
 	__asm {
 		pushf;
